requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
supabase==2.4.0
python-dateutil==2.8.2
gspread==6.0.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openpyxl==3.1.2
XlsxWriter==3.2.0
plotly==5.19.0
python-dotenv==1.0.1
uuid-utils==0.9.0
httpx>=0.28.0
